+++ routes/cp.php
@@ -2,7 +2,8 @@
 
 use Rias\StatamicRedirect\Controllers\Api\RedirectController as ApiRedirectController;
 use Rias\StatamicRedirect\Controllers\DashboardController;
-use Rias\StatamicRedirect\Controllers\Api\ErrorController;
+use Rias\StatamicRedirect\Controllers\Api\ErrorController as ApiErrorController;
+use Rias\StatamicRedirect\Controllers\ErrorController;
 use Rias\StatamicRedirect\Controllers\ExportController;
 use Rias\StatamicRedirect\Controllers\ImportRedirectsController;
 use Rias\StatamicRedirect\Controllers\RedirectController;
@@ -10,8 +11,14 @@ use Rias\StatamicRedirect\Controllers\RedirectController;
 Route::get('redirect/dashboard', '\\'. DashboardController::class)->name('redirect.index');
 Route::get('redirect/export/{type}', '\\'. ExportController::class)->name('redirect.export');
 
-Route::get('redirect/api/errors', ['\\' . ErrorController::class, 'index'])->name('redirect.api.errors.index');
+Route::get('redirect/api/errors', ['\\' . ApiErrorController::class, 'index'])->name('redirect.api.errors.index');
 Route::get('redirect/api/redirects', ['\\' . ApiRedirectController::class, 'index'])->name('redirect.api.redirects.index');
+
+Route::prefix('redirect/errors')->group(function () {
+    Route::get('clear', ['\\' . ErrorController::class, 'clearAll'])->name('redirect.api.errors.clear');
+    Route::get('/{id}', ['\\' . ErrorController::class, 'show'])->name('redirect.errors.show');
+    Route::get('/{id}/delete', ['\\' . ErrorController::class, 'delete'])->name('redirect.errors.delete');
+});
 
 Route::prefix('redirect/redirects')->group(function () {
     Route::get('/', ['\\' . RedirectController::class, 'index'])->name('redirect.redirects.index');