+++ src/RedirectServiceProvider.php
@@ -13,6 +13,7 @@ use Rias\StatamicRedirect\Listeners\CreateRedirect;
 use Rias\StatamicRedirect\Stache\Errors\ErrorStore;
 use Rias\StatamicRedirect\Stache\Redirects\RedirectStore;
 use Rias\StatamicRedirect\UpdateScripts\AddHitsCount;
+use Rias\StatamicRedirect\UpdateScripts\ClearErrors;
 use Statamic\Events\EntrySaved;
 use Statamic\Events\EntrySaving;
 use Statamic\Facades\CP\Nav;
@@ -26,6 +27,7 @@ class RedirectServiceProvider extends AddonServiceProvider
 {
     protected $updateScripts = [
         AddHitsCount::class,
+        ClearErrors::class,
     ];
 
     protected $scripts = [