+++ src/Stache/Errors/ErrorStore.php
@@ -3,9 +3,16 @@
 namespace Rias\StatamicRedirect\Stache\Errors;
 
 use Exception;
+use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Str;
 use Rias\StatamicRedirect\Facades\Error;
+use Statamic\Facades\File;
+use Statamic\Facades\Path;
+use Statamic\Facades\Stache;
 use Statamic\Facades\YAML;
+use Statamic\Stache\Exceptions\DuplicateKeyException;
 use Statamic\Stache\Stores\BasicStore;
+use Statamic\Stache\Traverser;
 use Symfony\Component\Finder\SplFileInfo;
 
 class ErrorStore extends BasicStore
@@ -17,6 +24,13 @@ class ErrorStore extends BasicStore
     public function key()
     {
         return 'errors';
+    }
+
+    public function getItemFilter(SplFileInfo $file)
+    {
+        $filename = str_after(Path::tidy($file->getPathName()), $this->directory);
+
+        return !Str::contains($filename, '_hits') && $file->getExtension() === 'yaml';
     }
 
     public function makeItemFromFile($path, $contents)
@@ -41,7 +55,6 @@ class ErrorStore extends BasicStore
             ->url(array_pull($data, 'url'))
             ->handled(array_pull($data, 'handled'))
             ->handledDestination(array_pull($data, 'handledDestination'))
-            ->hits(array_pull($data, 'hits'))
             ->hitsCount(array_pull($data, 'hitsCount'))
             ->initialPath($path);
 
@@ -50,10 +63,5 @@ class ErrorStore extends BasicStore
         }
 
         return $error;
-    }
-
-    public function filter(SplFileInfo $file)
-    {
-        return $file->getExtension() === 'yaml';
     }
 }